requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
attrs==24.2.0
bitsandbytes==0.44.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
coloredlogs==15.0.1
datasets==3.0.1
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
evaluate==0.4.3
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml_dtypes==0.5.0
mlc-ai-nightly-cpu==0.18.dev225
mlc-ai-nightly-cu122==0.18.dev225
mlc-llm-nightly-cpu==0.18.dev57
mlc-llm-nightly-cu122==0.18.dev57
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime==1.19.2
onnxruntime-genai-cuda==0.4.0
onnxruntime-gpu==1.19.2
openai==1.52.0
optimum==1.23.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pfzy==0.3.4
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
sympy==1.13.3
tensorrt==10.5.0
tensorrt-cu12==10.5.0
tensorrt-cu12-bindings==10.5.0
tensorrt-cu12-libs==10.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.9
tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
wheel==0.44.0
xxhash==3.5.0
yarl==1.15.2
